+++ src/features/gathering/ui/GatheringDetailButton.tsx
